refactor: split import_params, fix polyfill_verification Miri panic
import_params.rs: Pure import_intrinsics tests (6 tests) return to
archmage as import_intrinsics.rs. import_magetypes and combined tests
(14 tests) stay in magetypes as import_params.rs.

polyfill_verification.rs: Replace .expect("SSE4.1 required") with
early return when summon() returns None. Miri doesn't support SIMD
feature detection, so summon() returns None and the expect panicked.
