+++ magetypes/tests/polyfill_verification.rs
@@ -40,7 +40,12 @@ fn arrays_close(a: &[f32; 8], b: &[f32; 8], tolerance: f32) -> bool {
 
 #[test]
 fn verify_add() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => {
@@ -72,7 +77,12 @@ fn verify_add() {
 
 #[test]
 fn verify_sub() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -95,7 +105,12 @@ fn verify_sub() {
 
 #[test]
 fn verify_mul() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -118,7 +133,12 @@ fn verify_mul() {
 
 #[test]
 fn verify_div() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -141,7 +161,12 @@ fn verify_div() {
 
 #[test]
 fn verify_neg() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -166,7 +191,12 @@ fn verify_neg() {
 
 #[test]
 fn verify_min() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -192,7 +222,12 @@ fn verify_min() {
 
 #[test]
 fn verify_max() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -218,7 +253,12 @@ fn verify_max() {
 
 #[test]
 fn verify_abs() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -239,7 +279,12 @@ fn verify_abs() {
 
 #[test]
 fn verify_sqrt() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -260,7 +305,12 @@ fn verify_sqrt() {
 
 #[test]
 fn verify_floor() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -281,7 +331,12 @@ fn verify_floor() {
 
 #[test]
 fn verify_ceil() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -302,7 +357,12 @@ fn verify_ceil() {
 
 #[test]
 fn verify_round() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -327,7 +387,12 @@ fn verify_round() {
 
 #[test]
 fn verify_reduce_add() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -348,7 +413,12 @@ fn verify_reduce_add() {
 
 #[test]
 fn verify_reduce_max() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -369,7 +439,12 @@ fn verify_reduce_max() {
 
 #[test]
 fn verify_reduce_min() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -394,7 +469,12 @@ fn verify_reduce_min() {
 
 #[test]
 fn verify_edge_cases_infinity() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -442,7 +522,12 @@ fn verify_edge_cases_infinity() {
 
 #[test]
 fn verify_edge_cases_denormals() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -479,7 +564,12 @@ fn verify_edge_cases_denormals() {
 
 #[test]
 fn verify_mul_add() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -517,7 +607,12 @@ fn verify_mul_add() {
 
 #[test]
 fn verify_load_store_roundtrip() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
 
     let original = [1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0];
     let v = poly::f32x8::load(sse, &original);
@@ -528,7 +623,12 @@ fn verify_load_store_roundtrip() {
 
 #[test]
 fn verify_splat() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -548,7 +648,12 @@ fn verify_splat() {
 
 #[test]
 fn verify_zero() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
@@ -570,7 +675,12 @@ fn verify_zero() {
 
 #[test]
 fn verify_clamp() {
-    let sse = X64V3Token::summon().expect("SSE4.1 required");
+    let sse = {
+        let Some(t) = X64V3Token::summon() else {
+            return;
+        };
+        t
+    };
     let avx = match X64V3Token::summon() {
         Some(t) => t,
         None => return,
